running version of simulation
- add NetworkHandle::has_dep_agents() to check if a simulation interval
  has departure agents and apply it to simulation;
- change the argument of LinkQueue::to_interval from unsigned short
  to double as minute can be a floating number;
- add DemandPeriod::to_minutes() and apply it to setup_time() to get
  the time in minutes;
- update Agent::get_time_sequence() to get valid simulation interval only;
- add static_cast<double> to make sure NetworkHandle::cast_interval_to_minute_double()
  output proper double value;
- optimize minute calculation via % in NetworkHandle::get_time_stamp();
- update NetworkHandle::output_trajectories() to make sure the format is
  correct.

+++ include/supply.h
@@ -1434,7 +1434,7 @@ private:
         return c2 + residual;
     }
 
-    size_type to_interval(unsigned short m) const
+    size_type to_interval(double m) const
     {
         return std::ceil(m * SECONDS_IN_MINUTE / res);
     }
